% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_corr.R
\name{exp_corr}
\alias{exp_corr}
\title{Expected Local Ancestry Correlation}
\usage{
exp_corr(theta, g, props, k1, k2)
}
\arguments{
\item{theta}{recombination fraction between pair of loci}

\item{g}{generations since admixture}

\item{props}{data frame (n x K) of admixture proportions}

\item{k1, k2}{component indices}
}
\value{
A single number indicating the correlation of local ancestry.
}
\description{
Calculate expected correlation in local ancestry components at
a pair of loci, based on Lemma 1 in Grinde et al. (TBD).
}
\examples{
exp_corr(theta = 0.2, g = 6, props = example_props, k1 = 1, k2 = 1)
exp_corr(theta = 0.2, g = 6, props = example_props, k1 = 1, k2 = 2)

}
