% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ab.R
\name{get_ab}
\alias{get_ab}
\title{Helper Function: Calculate Scalars}
\usage{
get_ab(deltas, g)
}
\arguments{
\item{deltas}{vector of distances between consecutive markers}

\item{g}{the generations since admixture.}
}
\value{
List of two vectors with two sets of scalars.
}
\description{
Helper function for test statistic simulation.
Calculate scalars needed for test stat simulation.
}
