% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deltas.R
\name{get_deltas}
\alias{get_deltas}
\title{Pairwise distances between markers}
\usage{
get_deltas(map)
}
\arguments{
\item{map}{data frame with m rows and 2 columns ('chr' containing chromosome number and 'cM' containing genetic position in centimorgans), where m = no. markers}
}
\value{
Numeric vector (length m-1) of distances between markers.
}
\description{
Helper function for test statistic simulation.
Get distances between consecutive markers.
}
