% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_L.R
\name{get_L}
\alias{get_L}
\title{Expected correlation: one locus, square root}
\usage{
get_L(ep)
}
\arguments{
\item{ep}{vector of length K (K = no. ancestral pops) containg expected/average admixture proportions}
}
\value{
Matrix with K rows and K-1 columns.
}
\description{
Helper function for test statistic simulation.
Square root of expected correlation matrix at one locus.
}
