% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Expected correlation: one locus, all ancestries}
\usage{
get_sigma(ep)
}
\arguments{
\item{ep}{vector of length K (K = no. ancestral pops) containg expected/average admixture proportions}
}
\value{
Correlation matrix.
}
\description{
Helper function for test statistic simulation.
Calculate expected correlation at single locus
for all pairs of ancestries.
}
