% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma_ij.R
\name{get_sigma_ij}
\alias{get_sigma_ij}
\title{Expected correlation: one locus, one pair of ancestries}
\usage{
get_sigma_ij(ep1, ep2)
}
\arguments{
\item{ep1, ep2}{expected (average) admixture proportion for two ancestral populations}
}
\value{
Single number (the correlation).
}
\description{
Helper function for test statistic simulation.
Calculate expected correlation at single locus
for a single pair of ancestries.
}
