% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_approx.R
\name{OU_approx}
\alias{OU_approx}
\title{Ornstein Uhlenbeck Approximation to Family-Wise Error Rate}
\usage{
OU_approx(z, beta, Delta, length, chr, center, test)
}
\arguments{
\item{z}{the critical value.}

\item{beta}{the OU covariance parameter.}

\item{Delta}{the spacing between markers.}

\item{length}{the total length of the region.}

\item{chr}{the number of chromosomes.}

\item{center}{the level for family-wise error rate.}

\item{test}{the type of test (\code{"two-sided"} or \code{"one-sided"}).}
}
\value{
A single number indicating the distance between the family-wise error rate
    using critical value \code{z} and the desired level (\code{center}).
}
\description{
Analytic approximation to the family-wise error rate for test statistics
arising from Ornstein-Uhlenbeck process. Approximation courtesy of:
Siegmund and Yakir (2007). The statistics of gene mapping.
}
\examples{
OU_approx(4.277817, 0.01*6, 0.2, 3500, 22, 0.05, "two-sided")

}
