% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_alpha.R
\name{upper_alpha}
\alias{upper_alpha}
\title{Helper Function: Calculate Quantile}
\usage{
upper_alpha(t, alpha)
}
\arguments{
\item{t}{vector of test statistics}

\item{alpha}{the level for family-wise error rate control; default is 0.05.}
}
\value{
Single number.
}
\description{
Helper function for test statistic simulation.
Get upper alpha quantile of vector.
}
